refactor(domain/저장하는): null 반환 제거
